Crate risq

Rust Implementation of Semistochastic Quantum chemistry (RISQ)

A library for performing electronic structure calculations using advanced configuration interaction methods including Heat-bath CI (HCI) and Semistochastic HCI (SHCI).

Features

Heat-bath Configuration Interaction (HCI)

Semistochastic Epstein-Nesbet Perturbation Theory

Dynamic Semistochastic Full CI Quantum Monte Carlo (DS-FCIQMC)

Efficient determinant representation using bitstrings

Davidson eigenvalue solver with preconditioning

Alias sampling for importance sampling

Usage

use risq::context::RisqContext;
use risq::wf::init_var_wf;
use risq::var::variational;
use risq::pt::perturbative;

// Initialize calculation context
let mut context = RisqContext::from_files("in.json", "FCIDUMP")?;

// Set up wavefunction
let mut wavefunction = init_var_wf(&context)?;

// Run variational calculation
let var_energy = variational(&mut context, &mut wavefunction)?;

// Run perturbative correction
let pt2_energy = perturbative(&context, &wavefunction.wf)?;

println!("Total energy: {:.10}", var_energy + pt2_energy);

Re-exports

pub use context::RisqContext;

pub use error::RisqError;

pub use error::RisqResult;

Modules

config

Configuration Management (config)

context

Calculation Context (context)

error

Error Handling (error)

excite

Excitation Representation (excite)

ham

Hamiltonian Module (ham)

pt

Epstein-Nesbet Perturbation Theory Module (pt)

rng

Random Number Generation (rng)

semistoch

Semistochastic Methods (semistoch)

stoch

Stochastic Sampling Utilities (stoch)

temporary_wrappers

Temporary wrapper functions to interface between the new context-based architecture and the existing function signatures during refactoring. TODO: Remove this module once all functions are updated.

utils

General Utilities (utils)

var

Variational HCI Module (var)

wf

Wavefunction Module (wf)

Macros

risq_bail

Helper macros for common error patterns

risq_ensure

Constants

VERSION

Library version