risq

Module rng

Random Number Generation (rng)

This module provides a simple wrapper around a seeded random number generator (RNG) for use in stochastic parts of the calculations. Using a seeded RNG ensures reproducibility of stochastic results.

Structs

Rand

Wrapper struct holding the seeded random number generator instance.

Functions

init_rand

Initializes and returns a new Rand struct.