risq

Module semistoch

Semistochastic Methods (semistoch)

This module implements semistochastic approaches, primarily focusing on Epstein-Nesbet perturbation theory (ENPT2) combined with stochastic sampling. The goal is to capture the bulk of the PT2 correction deterministically while efficiently estimating the remaining contributions via sampling.

Functions

fast_stoch_enpt2

Implements a purely stochastic ENPT2 calculation

importance_sampled_semistoch_enpt2

Implements an importance-sampled semistochastic ENPT2 calculation.

new_semistoch_enpt2_dtm_diag_singles

old_semistoch_enpt2

Old algorithm (2017) for semistochastic ENPT2