risq

Module utils

General Utilities (utils)

This module provides various helper functions and submodules used throughout the risq crate.

Submodules:

bits: Functions for bit manipulation, primarily for working with determinant configurations.

display: Implementations of fmt::Display for custom data structures.

ints: Utilities related to handling molecular integrals and orbital indices.

read_input: Functionality for reading and parsing the in.json input file.

Modules

bits

Bit Manipulation Utilities (utils::bits)

display

Display Formatting Utilities (utils::display)

ints

Integral and Indexing Utilities (utils::ints)

read_input

Input File Reading (utils::read_input)