risq::utils

Module bits

Bit Manipulation Utilities (utils::bits)

This module provides helper functions and iterators for common bitwise operations, particularly useful for manipulating Slater determinant configurations (Config) represented as u128 bitstrings.

Functions

bit_pairs

Creates an iterator that yields unique pairs (i, j) where i < j and both bits i and j are set in n.

bits

Creates an iterator that yields the indices (0-based) of the set bits in n.

btest

Tests the b-th bit (0-based) of n. Returns true if the bit is 1, false otherwise. Equivalent to Fortran’s BTEST.

det_bits

Creates an iterator over all occupied spatial orbital indices in a determinant det, irrespective of spin. Combines the results of bits(det.up) and bits(det.dn).

elecs

Creates an iterator yielding all single occupied orbitals in det.

epairs

Creates an iterator yielding all pairs of occupied orbitals in det.

ibclr

Clears the b-th bit (0-based) of n to 0. Equivalent to Fortran’s IBCLR.

ibset

Sets the b-th bit (0-based) of n to 1. Equivalent to Fortran’s IBSET.

orbs

Creates an iterator yielding all single occupied orbitals and all pairs of occupied orbitals. Combines the output of epairs and elecs.

parity

Calculates the parity of the number of set bits in n using a fast parallel algorithm. Returns 1 if the number of set bits is even, -1 if it is odd. Useful for determining the sign factor when applying fermionic operators.

product_bits

Creates an iterator over all pairs (i, j) where i is an occupied alpha orbital and j is an occupied beta orbital in the determinant det. Represents opposite-spin electron pairs.

valence_elecs

Creates an iterator yielding single occupied valence orbitals in det.

valence_elecs_and_epairs

Creates an iterator yielding all single occupied valence orbitals and pairs of occupied valence orbitals within a determinant det.

valence_epairs

Creates an iterator yielding pairs of occupied valence orbitals in det.