risq::var::eigenvalues::algorithms

Module davidson

Davidson Diagonalization

The Davidson method is suitable for diagonal-dominant symmetric matrices, that are quite common in certain scientific problems like electronic structure. The Davidson method could be not practical for other kind of symmetric matrices.

The current implementation uses a general davidson algorithm, meaning that it compute all the requested eigenvalues simultaneusly using a variable size block approach. The family of Davidson algorithm only differ in the way that the correction vector is computed.

Available correction methods are:

DPR: Diagonal-Preconditioned-Residue

GJD: Generalized Jacobi Davidson

Structs

Davidson

Structure with the configuration data

DavidsonError