Module risq::var::off_diag

Data structure for holding off-diagonal Hamiltonian matrix elements and putting them into the SparseMat data structure for Davidson

Structs

OffDiagElemsNoHash

Off-diag elems data structure

Functions

add_el

Add an off-diagonal element H_{ij} to off_diag_elems i and j can be of any order

add_el_and_spin_flipped

Add an off-diagonal element H_{ij}, as well as its spin-flipped counterpart, to off_diag_elems i and j can be of any order