risq::var::off_diag

Function add_el

pub fn add_el(wf: &mut VarWf, ham: &Ham, i: usize, j: usize, elem: Option<f64>)

Computes (if necessary) and adds an off-diagonal Hamiltonian element H_ij to the sparse matrix representation stored in wf.sparse_ham.

Handles index ordering (i vs j) to store only the upper triangle. If elem is None, it computes H_ij using ham.ham_off_diag_no_excite. If the computed or provided element is non-zero, it calls add_off_diag_elem. Does nothing if i == j.

Arguments

wf: Mutable reference to the variational wavefunction, containing the sparse matrix.

ham: The Hamiltonian operator.

i, j: Indices of the two determinants in wf.wf.dets.

elem: Optionally pre-computed value of H_ij. If None, it will be computed.