risq::var::off_diag

Function add_el_and_spin_flipped

pub fn add_el_and_spin_flipped(wf: &mut VarWf, ham: &Ham, i: usize, j: usize)

Computes and adds both H_ij and its spin-flipped counterpart H_i'j' to the sparse matrix.

Assumes spin symmetry, where H_ij = H_i'j' if i' and j' are the spin-flips of i and j.

Computes elem = H_ij using ham.ham_off_diag_no_excite.

If elem is non-zero: a. Adds (j, elem) or (i, elem) to the sparse matrix via add_el. b. Finds the indices i_spin_flipped and j_spin_flipped corresponding to the spin-flipped configurations of determinants i and j using the wf.wf.inds map. c. Adds (j_spin_flipped, elem) or (i_spin_flipped, elem) to the sparse matrix via add_el. Does nothing if i == j.

Arguments

wf: Mutable reference to the variational wavefunction.

ham: The Hamiltonian operator.

i, j: Indices of the two determinants.

Panics

Panics if the spin-flipped configuration of determinant i or j is not found in wf.wf.inds.