risq::wf::eigensolver

Trait EigenSolver

pub trait EigenSolver {
    // Required methods
    fn solve(
        &mut self,
        wf: &mut VarWf,
        ham: &Ham,
        excite_gen: &ExciteGenerator,
        config: &EigenConfig,
    ) -> Result<EigenSolution, EigenSolverError>;
    fn is_converged(&self) -> bool;
    fn iterations(&self) -> usize;
    fn name(&self) -> &'static str;
}

Trait for eigenvalue solver algorithms

This trait abstracts different eigenvalue solving approaches like Davidson, Lanczos, or other iterative methods used in quantum chemistry.

Required Methods

fn solve( &mut self, wf: &mut VarWf, ham: &Ham, excite_gen: &ExciteGenerator, config: &EigenConfig, ) -> Result<EigenSolution, EigenSolverError>

Solve the eigenvalue problem for the given system

Arguments

wf - The variational wavefunction containing the system

ham - The Hamiltonian operator

excite_gen - Excitation generator

config - Solver configuration

Returns

Solution containing eigenvalues and eigenvectors, or error

fn is_converged(&self) -> bool

Check if the solver has converged

fn iterations(&self) -> usize

Get the number of iterations performed

fn name(&self) -> &'static str

Get solver name for debugging/logging