Module risq::wf::eps

Variational epsilon iterator (to attach to variational wf) Epsilon starts at the largest value that allows at least one double excitation from the initial wf, then drops by a factor of 2 every iteration until it reaches the target value set in the input file

Structs

Eps

Variational epsilon iterator

Functions

init_eps

Initialize epsilon iterator max_doub is the min of the largest symmetrical and largest asymmetrical double excitation magnitudes coming from the wavefunction Can’t just use excite_gen.max_(same/opp)_spin_doub because we want to only consider excitations coming from initial wf (usually HF det) We use this initial eps so that when we do excited states, there will be at least two closed shell and at least two open shell determinants