risq::wf::traits

Trait MatrixVectorProduct

pub trait MatrixVectorProduct {
    // Required method
    fn apply_hamiltonian(
        &self,
        ham: &Ham,
        excite_gen: &ExciteGenerator,
        config: &MatVecConfig,
    ) -> MatVecResult;

    // Provided methods
    fn apply_external(
        &self,
        ham: &Ham,
        excite_gen: &ExciteGenerator,
        eps: f64,
        singles_strategy: SinglesStrategy,
    ) -> MatVecResult { ... }
    fn apply_variational(
        &self,
        input_coeffs: &[f64],
        ham: &Ham,
        excite_gen: &ExciteGenerator,
        eps: f64,
    ) -> Vec<f64> { ... }
}

Trait for matrix-vector multiplication operations on wavefunctions

This trait unifies the various approx_matmul_* functions by providing a common interface for different matrix-vector product strategies.

Required Methods

fn apply_hamiltonian( &self, ham: &Ham, excite_gen: &ExciteGenerator, config: &MatVecConfig, ) -> MatVecResult

Apply the Hamiltonian to the wavefunction using specified configuration

Arguments

ham - The Hamiltonian operator

excite_gen - Excitation generator for determinant connections

config - Configuration specifying the operation parameters

Returns

Result of the matrix-vector multiplication based on the configuration

Provided Methods

fn apply_external( &self, ham: &Ham, excite_gen: &ExciteGenerator, eps: f64, singles_strategy: SinglesStrategy, ) -> MatVecResult

Convenience method for external perturbative space operations

fn apply_variational( &self, input_coeffs: &[f64], ham: &Ham, excite_gen: &ExciteGenerator, eps: f64, ) -> Vec<f64>

Convenience method for variational space operations