Crate risq

Rust Implementation of Semistochastic Quantum chemistry (risq) implements an efficient selected configuration interaction algorithm called Semistochastic Heat-bath Configuration Interaction (SHCI).

Modules

excite

Data structure that enables generating the most important excitations and importance sampling the rest

ham

Functions to compute Hamiltonian matrix elements

pt

Epstein-Nesbet perturbation theory

rng

Contains the random number generator and its seed

semistoch

Semistochastic methods - for now, just includes Epstein-Nesbet perturbation theory

stoch

General stochastic functions

utils

Misc useful utiliities

var

Variational stage of Heat-bath Configuration Interaction

wf

Variational wavefunction data structure Includes both the variational wf and the variational H Includes functions for initializing, printing, and adding new determinants

Functions

main