Module risq::excite

Data structure that enables generating the most important excitations and importance sampling the rest

Modules

init

Initialize sorted excitation arrays

iterator

Structs

Excite

Candidate excitation from one determinant to another. Contains the initial orbs init, target orb(s) target, absolute value of the excitation matrix element abs_h, and whether it is an alpha-spin excitation is_alpha (is_alpha = None for opposite-spin double excitations)

StoredExcite

Excitation information to be stored in ExciteGenerator. Contains only target and abs_h (since init and is_alpha are already known by the time StoredExcites are needed). Also contains sum_remaining_abs_h and sum_remaining_h_squared for importance-sampling the remaining excitations

Enums

Orbs

Generalized type that can either be a single spatial orbital or a pair of them