Module risq::utils::bits

Useful bitwise functions Bits(n) iterates over set bits in n, bit_pairs(n) iterates over pairs of set bits in n, plus functions for computing parity and getting and setting bits

Structs

BitPairs

Backend for bit_pairs()

BitPairsIntoIterator

Bits

Backend for bits()

BitsIntoIterator

Functions

bit_pairs

Iterate over pairs of set bits in a u128 Syntax: for (i, j) in bit_pairs(det: u128): loops over the unique pairs of set bits in det

bits

Iterate over set bits in a u128 Syntax: for i in bits(det: u128): loops over the set bits in det

btest

det_bits

Iterate over set bits in a Config

ibclr

ibset

parity