Module risq::wf

Variational wavefunction data structure Includes both the variational wf and the variational H Includes functions for initializing, printing, and adding new determinants

Modules

det

Determinant data structure Includes functions to generate an excited det, and compute its diagonal element quickly from the initial det’s diagonal element

eps

Variational epsilon iterator (to attach to variational wf) Epsilon starts at the largest value that allows at least one double excitation from the initial wf, then drops by a factor of 2 every iteration until it reaches the target value set in the input file

Structs

Wf

Variational wavefunction data structure

Functions

init_var_wf

Initialize variational wf to the HF det (only needs to be called once)